Renamed internal package gh -> graphql
