+++ internal/graphql/project_next_fields_test.go
@@ -1,4 +1,4 @@
-package gh
+package graphql
 
 import (
 	"testing"
